refactor: move validator status type and util to @lodestar/types
